Add all files to meet project specification
To block out the overall shape of the project which is described
in The Odin Project, the recipe pages have been added to the
project, as well as basic boilerplate HTML, an image and the
basic structure of each page.

This is also a test to see if I understood how to use VS Code to
write a commit message.